services/catalog/requirements.txt
flask==3.0.2
mysql-connector-python==9.0.0
pyjwt==2.8.0
flask-cors==4.0.0
